requirements/base.txt
amqp==5.2.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2023.10
cron-descriptor==1.4.3
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-celery-beat==2.5.0
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-debug==0.1.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.1
django-model-utils==4.3.1
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
exceptiongroup==1.2.0
flower==2.0.1
h11==0.14.0
hiredis==2.2.3
httptools==0.6.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
oauthlib==3.2.2
Pillow==10.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.4.0
rpds-py==0.18.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.24.0.post1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
whitenoise==6.6.0
